feat: Improve performance and refactor flight tracker
This commit introduces several improvements to the flight tracker application to enhance performance and code maintainability.

- Implemented bounding box queries to the OpenSky Network API, significantly reducing the amount of data fetched and processed by the client.
- Refactored the `App` component by extracting the settings panel and map/canvas logic into a separate component and a custom hook, respectively.
- Optimized the heatmap rendering logic to improve performance when the heatmap is enabled.
- Implemented a proxy for OpenSky API requests and configured the application to use it when the "Use Proxy" setting is enabled.
